+++ AlphaFS/Filesystem/FileInfo Class/FileInfo.Create.cs
@@ -33,9 +33,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public FileStream Create()
       {
-         return File.CreateFileStreamCore(Transaction, LongFullName, ExtendedFileAttributes.Normal, null,
-            FileMode.Create, FileAccess.ReadWrite, FileShare.None, NativeMethods.DefaultFileBufferSize,
-            PathFormat.LongFullPath);
+         return File.CreateFileStreamCore(Transaction, LongFullName, ExtendedFileAttributes.Normal, null, FileMode.Create, FileAccess.ReadWrite, FileShare.None, NativeMethods.DefaultFileBufferSize, PathFormat.LongFullPath);
       }
 
       #endregion // .NET